\docType{data}
\name{classifications}
\alias{classifications}
\title{RUCA classifications.}
\format{a data frame 34 observations of 7 variables.}
\source{
  Rural Health Research Center

  Rural Health Research Center
}
\description{
  Data frame containing the RUCA classifications.

  This data frames provides a mapping between zip codes and
  RUCA codes.
}
\keyword{datasets}

