\docType{package}
\name{ruca-package}
\alias{ruca}
\alias{ruca-package}
\title{Rural-Urban Commuting Area Codes}
\description{
  Rural-Urban Commuting Area Codes
}
\details{
  Given a postal code, will determine the urbanicity of
  that region based upon Rural Health Research Center's
  Rural-Urban Commuting Area Codes (RUCAs).
}
\author{
  \email{jason@bryer.org}
}
\keyword{package}
\keyword{ruca}
\keyword{urbanization}

